requirements.txt
# FloodLM Requirements
# Install with: conda create -n floodlm --file requirements.txt
# Or: pip install -r requirements.txt

# Core dependencies
python=3.10
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# PyTorch and Geometric
pytorch::pytorch::*=*cuda11*
pytorch::pytorch-cuda=11.8
pytorch::pytorch-geometric
pytorch::pyg

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Development
jupyter>=1.0.0
ipython>=8.0.0
pytest>=7.0.0
